+++ src/model/generate_model.jl
@@ -94,10 +94,10 @@ function generate_model(setup::Dict,inputs::Dict,OPTIMIZER::MOI.OptimizerWithAtt
 	EP[:eObj] = AffExpr(0.0)
 
 	create_empty_expression!(EP, :eGenerationByZone, (Z, T))
-	
+
 	# Energy losses related to technologies
 	create_empty_expression!(EP, :eELOSSByZone, Z)
-	
+
 	# Initialize Capacity Reserve Margin Expression
 	if setup["CapacityReserveMargin"] > 0
 		create_empty_expression!(EP, :eCapResMarBalance, (inputs["NCapacityReserveMargin"], T))
@@ -190,6 +190,11 @@ function generate_model(setup::Dict,inputs::Dict,OPTIMIZER::MOI.OptimizerWithAtt
 	end
 
 	# Policies
+
+	if setup["Reserves"] > 0
+		reserves_constraints!(EP, inputs)
+	end
+
 	# CO2 emissions limits
 	if setup["CO2Cap"] > 0
 		co2_cap!(EP, inputs, setup)