+++ src/model/resources/storage/storage_asymmetric.jl
@@ -42,35 +42,21 @@ Sets up variables and constraints specific to storage resources with asymmetric 
 """
 function storage_asymmetric_reserves!(EP::Model, inputs::Dict, setup::Dict)
 
-	dfGen = inputs["dfGen"]
 	T = inputs["T"]
-	CapacityReserveMargin = setup["CapacityReserveMargin"]
+	CapacityReserveMargin = setup["CapacityReserveMargin"] > 0
 
 	STOR_ASYMMETRIC = inputs["STOR_ASYMMETRIC"]
-
 	STOR_ASYM_REG = intersect(STOR_ASYMMETRIC, inputs["REG"]) # Set of asymmetric storage resources with REG reserves
-	STOR_ASYM_NO_REG = setdiff(STOR_ASYMMETRIC, STOR_ASYM_REG) # Set of asymmetric storage resources without REG reserves
 
-	if !isempty(STOR_ASYM_REG)
-		if CapacityReserveMargin > 0
-			# Storage units charging can charge faster to provide reserves down and charge slower to provide reserves up
-			# Maximum charging rate plus contribution to regulation down must be less than charge power rating
-			@constraint(EP, [y in STOR_ASYM_REG, t in 1:T], EP[:vCHARGE][y,t]+EP[:vREG_charge][y,t]+EP[:vCAPRES_charge][y,t] <= EP[:eTotalCapCharge][y])
-		else
-			# Storage units charging can charge faster to provide reserves down and charge slower to provide reserves up
-			# Maximum charging rate plus contribution to regulation down must be less than charge power rating
-			@constraint(EP, [y in STOR_ASYM_REG, t in 1:T], EP[:vCHARGE][y,t]+EP[:vREG_charge][y,t]<= EP[:eTotalCapCharge][y])
-		end
-	else
-		if CapacityReserveMargin > 0
-			# Storage units charging can charge faster to provide reserves down and charge slower to provide reserves up
-			# Maximum charging rate plus contribution to regulation down must be less than charge power rating
-			@constraint(EP, [y in STOR_ASYM_NO_REG, t in 1:T], EP[:vCHARGE][y,t]+EP[:vCAPRES_charge][y,t] <= EP[:eTotalCapCharge][y])
-		else
-			# Storage units charging can charge faster to provide reserves down and charge slower to provide reserves up
-			# Maximum charging rate plus contribution to regulation down must be less than charge power rating
-			@constraint(EP, [y in STOR_ASYM_NO_REG, t in 1:T], EP[:vCHARGE][y,t]<= EP[:eTotalCapCharge][y])
-		end
-	end
+    vCHARGE = EP[:vCHARGE]
+    vREG_charge = EP[:vREG_charge]
+    eTotalCapCharge = EP[:eTotalCapCharge]
 
+    expr = extract_time_series_to_expression(vCHARGE, STOR_ASYMMETRIC)
+    add_similar_to_expression!(expr[STOR_ASYM_REG, :], vREG_charge[STOR_ASYM_REG, :])
+    if CapacityReserveMargin
+        vCAPRES_charge = EP[:vCAPRES_charge]
+        add_similar_to_expression!(expr[STOR_ASYMMETRIC, :], vCAPRES_charge[STOR_ASYMMETRIC, :])
+    end
+    @constraint(EP, [y in STOR_ASYMMETRIC, t in 1:T], expr[y, t] <= eTotalCapCharge[y])
 end