+++ src/write_outputs/write_energy_revenue.jl
@@ -11,14 +11,15 @@ function write_energy_revenue(path::AbstractString, inputs::Dict, setup::Dict, E
 	NONFLEX = setdiff(collect(1:G), FLEX)
 	dfEnergyRevenue = DataFrame(Region = dfGen.region, Resource = inputs["RESOURCES"], Zone = dfGen.Zone, Cluster = dfGen.cluster, AnnualSum = Array{Float64}(undef, G),)
 	energyrevenue = zeros(G, T)
-	energyrevenue[NONFLEX, :] = value.(EP[:vP][NONFLEX, :]) .* transpose(dual.(EP[:cPowerBalance]) ./ inputs["omega"])[dfGen[NONFLEX, :Zone], :]
+    price = locational_marginal_price(EP, inputs, setup)
+    energyrevenue[NONFLEX, :] = value.(EP[:vP][NONFLEX, :]) .* transpose(price)[dfGen[NONFLEX, :Zone], :]
 	if !isempty(FLEX)
-		energyrevenue[FLEX, :] = value.(EP[:vCHARGE_FLEX][FLEX, :]).data .* transpose(dual.(EP[:cPowerBalance]) ./ inputs["omega"])[dfGen[FLEX, :Zone], :]
+		energyrevenue[FLEX, :] = value.(EP[:vCHARGE_FLEX][FLEX, :]).data .* transpose(price)[dfGen[FLEX, :Zone], :]
 	end
 	if setup["ParameterScale"] == 1
-		energyrevenue *= ModelScalingFactor^2
+		energyrevenue *= ModelScalingFactor
 	end
 	dfEnergyRevenue.AnnualSum .= energyrevenue * inputs["omega"]
-	CSV.write(joinpath(path, "EnergyRevenue.csv"), dfEnergyRevenue)
+	write_simple_csv(joinpath(path, "EnergyRevenue.csv"), dfEnergyRevenue)
 	return dfEnergyRevenue
 end